update etuff test data and test suite
