+++ src/test/java/edu/ucar/unidata/rosetta/converters/known/TestTagUniversalFileFormatOneLocManyOb.java
@@ -13,7 +13,7 @@ import java.io.File;
 import java.io.IOException;
 import java.util.List;
 
-import edu.ucar.unidata.rosetta.converters.known.TagUniversalFileFormat;
+import edu.ucar.unidata.rosetta.converters.known.etuff.TagUniversalFileFormat;
 import edu.ucar.unidata.rosetta.util.test.util.TestUtils;
 import ucar.ma2.Array;
 import ucar.ma2.InvalidRangeException;
@@ -34,34 +34,32 @@ public class TestTagUniversalFileFormatOneLocManyOb {
     private static String etuffDir = String
             .join(File.separator, "conversions", "TagUniversalFileFormat");
     private static String etuffFileTld = TestUtils.getTestDataDirStr() + etuffDir;
-    //private static String etuffFile = String.join(File.separator, etuffFileTld, "eTUFF-tuna-590051.gz");
+
     private static String etuffFile = String
-            .join(File.separator, etuffFileTld, "TagDataFlatFileExample.gz");
+            .join(File.separator, etuffFileTld, "eTUFF-tuna-590051-small.txt");
     private static String etuffFileNc = String
-            .join(File.separator, etuffFileTld, "eTUFF-tuna-590051.nc");
+            .join(File.separator, etuffFileTld, "eTUFF-tuna-590051-small.nc");
     private static NetcdfDataset ncd;
+
+    private static int numRecords = 1594;
 
     private TagUniversalFileFormat etuffConvertor = new TagUniversalFileFormat();
 
     @BeforeClass
-    public static void sanityCheckAndOutputRead() {
+    public static void sanityCheckAndOutputRead() throws IOException, InvalidRangeException {
         TagUniversalFileFormat converter = new TagUniversalFileFormat();
         converter.setMatchupOneLocOneOb(false);
+        // won't always have netCDF-C lib installed where running tests
+        converter.setUseNetcdf4(false);
+
         // check to make sure no global metadata has been set
         assertEquals(converter.getGlobalMetadata().size(), 0);
         converter.parse(etuffFile);
         // check to make sure global metadata was parsed
         //assertEquals(converter.getGlobalMetadata().size(), 39);
 
-        try {
-            String ncfile = converter.convert(etuffFileNc);
-            ncd = NetcdfDataset.openDataset(ncfile);
-
-        } catch (InvalidRangeException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        String ncfile = converter.convert(etuffFileNc);
+        ncd = NetcdfDataset.openDataset(ncfile);
     }
 
     @AfterClass
@@ -76,7 +74,7 @@ public class TestTagUniversalFileFormatOneLocManyOb {
         List<CoordinateAxis> cas = ncd.getCoordinateAxes();
         assertEquals(cas.size(), 5);
         Dimension timeDim = ncd.findDimension("time");
-        assertEquals(timeDim.getLength(), 43201);
+        assertEquals(numRecords, timeDim.getLength());
     }
 
     @Test
@@ -89,7 +87,7 @@ public class TestTagUniversalFileFormatOneLocManyOb {
                 List<Dimension> dims = var.getDimensions();
                 assertEquals(dims.size(), 1);
                 Dimension timeDim = dims.get(0);
-                assertEquals(timeDim.getLength(), 43201);
+                assertEquals(numRecords, timeDim.getLength());
             }
         }
     }
@@ -97,31 +95,38 @@ public class TestTagUniversalFileFormatOneLocManyOb {
     @Test
     public void testNoMatchup() throws IOException {
         // for this date and ob:
-        //   2005-07-01 0:06:00,6,14.60,"temperature","Celsius"
+        // "2005-04-15 21:33:00",6,23.30,"temperature","Celsius"
         // there is no lat/lon matchup
-        // this tests makes sure it is matches with a location, as
+        // this tests makes sure it is matched with a location, as
         // we are doing a many ob to one loc matchup
-        CalendarDate findTimeCd = CalendarDate.parseISOformat("gregorian", "2005-07-01 0:06:00");
+        CalendarDate findTimeCd = CalendarDate.parseISOformat("gregorian", "2005-04-15 21:33:00");
         long findTime = findTimeCd.toDate().getTime() / 1000;
         Variable time = ncd.findVariable("time");
         Array timeVals = time.read();
 
         boolean found = false;
+        int foundIndex = 0;
         for (int i = 0; i < timeVals.getSize(); i++) {
             long checkTime = timeVals.getLong(i);
             if (checkTime == findTime) {
                 found = true;
+                foundIndex = i;
             }
         }
         assertTrue(found);
+
+        // check that temperature value is as expected
+        Variable temp = ncd.findVariable("temperature");
+        Array tempVals = temp.read();
+        assertEquals(23.30, tempVals.getFloat(foundIndex), 0.001);
     }
 
     @Test
     public void testMatchup() throws IOException {
         // for this date and ob:
-        //   2005-07-10 0:00:00,6,15.10,"temperature","Celsius"
+        //  "2005-04-16 00:00:00",6,21.50,"temperature","Celsius"
         // there is a lat/lon matchup
-        CalendarDate findTimeCd = CalendarDate.parseISOformat("gregorian", "2005-07-10 0:00:00");
+        CalendarDate findTimeCd = CalendarDate.parseISOformat("gregorian", "2005-04-16 00:00:00");
         long findTime = findTimeCd.toDate().getTime() / 1000;
         Variable time = ncd.findVariable("time");
         Array timeVals = time.read();
@@ -139,6 +144,6 @@ public class TestTagUniversalFileFormatOneLocManyOb {
         // check that temperature value is as expected
         Variable temp = ncd.findVariable("temperature");
         Array tempVals = temp.read();
-        assertEquals(tempVals.getFloat(foundIndex), 15.10, 0.001);
+        assertEquals(21.50, tempVals.getFloat(foundIndex), 0.001);
     }
 }
